add conditions for schema migrations

+++ database/migrations/2024_06_04_114604_create_local_market_order_has_inventories_table.php
@@ -13,18 +13,20 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('local_market_order_has_inventories', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('local_market_order_id');
-            $table->foreign('local_market_order_id')->references('id')->on('local_market_orders')->onDelete('cascade');
-            $table->unsignedBigInteger('local_market_inventory_id');
-            $table->foreign('local_market_inventory_id', 'fk_lm_inventory')->references('id')->on('local_market_inventories');
-            $table->integer('quantity');
-            $table->decimal('price', 64, 0);
-            $table->unsignedBigInteger('supplier_id');
-            $table->foreign('supplier_id')->references('id')->on('companies');
-            $table->timestamps();
-        });
+        if (! Schema::hasTable('local_market_order_has_inventories')) {
+            Schema::create('local_market_order_has_inventories', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('local_market_order_id');
+                $table->foreign('local_market_order_id')->references('id')->on('local_market_orders')->onDelete('cascade');
+                $table->unsignedBigInteger('local_market_inventory_id');
+                $table->foreign('local_market_inventory_id', 'fk_lm_inventory')->references('id')->on('local_market_inventories');
+                $table->integer('quantity');
+                $table->decimal('price', 64, 0);
+                $table->unsignedBigInteger('supplier_id');
+                $table->foreign('supplier_id')->references('id')->on('companies');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
